pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "bert-score>=0.3.13",
    "litellm>=1.80.12",
    "loguru>=0.7.3",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
